requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
requests>=2.31.0
httpx>=0.24.0
openai>=1.0.0